Prototype the drawing from container, dispensing into container, and transferring the liquid between containers.
Pipettes from one container to another. Robot only knows how to draw given volume from source and dispense all of it into the destination. Partial dispensing (i.e. dispensing of not all the volume held in the pipette) is not implemented yet.

Tested on one stock 20mL bottle (`bottle6`) and one 54-vial wellplate. Well number 45 is deleted because this corner of the well plate cannot be reached by the delta robot.