Speed up xy motions and end-of-motion status checks
This eliminates the delays at the end of each xy motion that were previously caused by polling the GRBL for status until the status is "Idle". Now, instead of polling GRBL for status updates, I predict the time that the particular motion will take based on acceleration and max speed -- and then I simply sleep for this time.