Change logging priorities of the printMSG calls and set INFO=1, do not show
Before this commit, messages (commands) to Zeus were displayed in yellow "WARNING" color, which does not make sense, because sending a command does not constitute a "warning": it's a standard operation, nothing to worry about. "Errors" that are equal to "None" are not displayed anymore in red color, as these are not errors per se.

Message displayed in the error parsing method starting with "ErrorString =" now reads "Message from Zeus: " because that's what it actually is. Preferably, this INFO message should be invoked in a more appropriate place of the code, not in the error parsing method.