Process the `multicomp-reactions/2023-06-19-run01`, find the mistakes and outliers to be remeasured
Some of the mistakes were marked by setting the `is_outlier` flag to 1 in the `run_structure.csv` files after the experiment (this reflected the experimental notes in respective `README.md` files in each run. These mistakes were included into the conditions to be remeasured once: see script `list_known_outliers.py`

Some outliers were identified manually by inspecting the data cube with `animated_viewer_interp.py` and filling the list of indices in the `manually_mark_outliers.py` script. These outliers are to be repeated twice.

The `request_outliers.py` file concatenates the lists of known (to be repeated once) and manual outliers (to be repeated twice), shuffles the list and saves to the Excel workbook file used by the pipetter.