Locate condition by operation datetime and plate id (barcode)
Added the following method and its tests. Given the dataframe of conditions, the time of operation and the barcode of the plate, method finds the condition that correspond to each container on which the operation was performed. For example, if the operation is measurement of spectra of containers in this plate, then it will find indices of conditions that correspond to each measured container (vial, well) of the plate. The logic of this method is: for each container of a given plate (by plate barcode), the method finds the latest condition whose timestamp precedes the input timestamp.