Add nanodrop compatibility to UV-VIS spectral processing code
Now that load spectra from the files are checking if the 'nanodrop' is in the file path. If it is, new code is used to load the nanodrop file (one file per plate). If it is not, the old code is used to load CRAIC folder (one folder per plate).