Use Cary Agilent data for reference spectra of HRP01 and dm37
Before this commit, the reference spectra used in the calibrator were from nanodrop. In this commit, the calibrator can use Agilent Cary spectra as references instead. The calibration curve is still constructed using the nanodrop spectra, only the reference spectrum is from Cary. So far, I've measured spectra on Cary for the HRP01 and dm37. Similarly measured spectra for other components were not processed in this commit yet.