Use Cary Agilent data for almost all reference spectra and save residuals by calibrator
Before this commit, the reference spectra used in the calibrator were from nanodrop, with the exception of HRP01 and dm37. In this commit, the spectra measured by Agilent Cary are used as references for all the components of Hantzsch instead. The calibration curve is still constructed using the nanodrop spectra, only the reference spectrum is from Cary.

Calibrator saves residuals from each fit in order to later construct a model of nanodrop's errorbars (in terms of absorbance). Add an option to not save calibration data when running `calibrator.py`